reviewed applyAndAcceptloan

+++ banking/src/main/java/com/revature/p1/banking/DAO/LoanDAO.java
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LoanDAO extends JpaRepository<Loan, Long> { }
+public interface LoanDAO extends JpaRepository<Loan, Integer> { }
